fix(training): move capacity helper into rooms module to satisfy boundary checks
getTrainingRoomCapacity lives in src/modules/rooms/utils, matching the
modules/map/utils precedent; core must not import from modules.
