+++ frontend/src/modules/rooms/utils/room.ts
